Change divisor? to speed up some of the solutions
In particular the solution for problem 23 will be a little bit faster.
The divisor? function now expects just one divisor to check, not a
variable list of arguments.

The 'old' divisor? function is now called some-divisor and will return
the first divisor (if any, else nil) instead of true or false.